Handle invalid iterator access in doPoll
There is a probablity that if a new drive comes up or existing
drive got removed the loop over drives in doPoll method can
cause invalid iterator access. Hence, keep a copy of drives
for the loop and use the loop over copied loop.

Tested:
frequently restarted mctpd service and verified the insertion
and removal of drives. verified the NVMe service was running.

Change-Id: I9d0046d7fffe6ee32f38ec9af3f1d04244f47387
Signed-off-by: Amit Kumar Tiwari <amitx.kumar.tiwari@intel.com>