Trigger the poll of device
Add a mechanism to trigger the poll when new drive is added
and stop the same when there is no drive connected. No need
of HS poll task to be alive if there is no drive connected.
And also it should be triggered if the first drive gets connected.

Tested:
Verified drive is detected and polling is started.
Verified poll is stopped when all drives are removed.

Change-Id: I4a66e6393673c776a4272d69014e4c80c99ef190
Signed-off-by: Amit Kumar Tiwari <amitx.kumar.tiwari@intel.com>